Removed embedded database, now connects to the local db. Refactored FrameController and LexUnitController. Using Hibernate.Initialize to instantiate. Using LightLU instead of LexUnit objects (only id, name and frame). Fixed bug in PatternEntry when comparing valence patterns. Fixed bugs in TabController. Fixed bugs in instances variable types (FrameRelation, Label, Sentence, StatusType). Removed dynamic loading of tabs.

+++ src/main/java/jp/keio/jfn/wat/domain/StatusType.java
@@ -24,7 +24,7 @@ public class StatusType implements Serializable {
 
 	private String name;
 
-	private byte rank;
+	private int rank;
 
 	//bi-directional many-to-one association to Status
 	@OneToMany(mappedBy="statusType")
@@ -65,11 +65,11 @@ public class StatusType implements Serializable {
 		this.name = name;
 	}
 
-	public byte getRank() {
+	public int getRank() {
 		return this.rank;
 	}
 
-	public void setRank(byte rank) {
+	public void setRank(int rank) {
 		this.rank = rank;
 	}
